# Server/Source/server/server-web/src/main/resources/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "pricediff Rest Services",
    "version" : "v1",
    "title" : "pricediff",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "email" : "pricediff@hacktm.de"
    }
  },
  "host" : "[domain]:[port]",
  "basePath" : "/pricediff/rs",
  "tags" : [ {
    "name" : "rest",
    "description" : "Products resource"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/rest/add" : {
      "get" : {
        "tags" : [ "rest" ],
        "summary" : "Add product",
        "description" : "",
        "operationId" : "addProduct",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "gpsPosition",
          "in" : "query",
          "description" : "gps position.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "barcode",
          "in" : "query",
          "description" : "barcode.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "name",
          "in" : "query",
          "description" : "name.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "userId",
          "in" : "query",
          "description" : "userId.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "price",
          "in" : "query",
          "description" : "price.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "category",
          "in" : "query",
          "description" : "category.",
          "required" : true,
          "type" : "string",
          "enum" : [ "LACTATE", "FRUCTE", "CARNE", "DULCIURI" ]
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ProductMdl"
            }
          }
        }
      }
    },
    "/rest/addProfile" : {
      "get" : {
        "tags" : [ "rest" ],
        "summary" : "Add profile",
        "description" : "",
        "operationId" : "addProfile",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "username.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "password.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "email",
          "in" : "query",
          "description" : "email.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ProfileMdl"
            }
          }
        }
      }
    },
    "/rest/deal" : {
      "get" : {
        "tags" : [ "rest" ],
        "summary" : "Find best deal",
        "description" : "",
        "operationId" : "findBestDeal",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "list",
          "in" : "query",
          "description" : "list of product IDs",
          "required" : true,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ProductResponseMdl"
              }
            }
          }
        }
      }
    },
    "/rest/find" : {
      "get" : {
        "tags" : [ "rest" ],
        "summary" : "Find product",
        "description" : "",
        "operationId" : "findProduct",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "category",
          "in" : "query",
          "description" : "search category.",
          "required" : true,
          "type" : "string",
          "enum" : [ "LACTATE", "FRUCTE", "CARNE", "DULCIURI" ]
        }, {
          "name" : "query",
          "in" : "query",
          "description" : "search query.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ProductMdl"
              }
            }
          }
        }
      }
    },
    "/rest/resp" : {
      "get" : {
        "tags" : [ "rest" ],
        "summary" : "Find resp",
        "description" : "",
        "operationId" : "productResponse",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ProductResponseMdl"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "PriceMdl" : {
      "type" : "object",
      "properties" : {
        "price" : {
          "type" : "number",
          "format" : "double"
        },
        "product" : {
          "$ref" : "#/definitions/ProductMdl"
        },
        "profile" : {
          "$ref" : "#/definitions/ProfileMdl"
        },
        "createdAt" : {
          "type" : "string",
          "format" : "date-time"
        },
        "store" : {
          "type" : "string",
          "enum" : [ "KAUFLAND_SAGULUI", "KAUFLAND_LAZAR", "PROFI_COMPLEX" ]
        }
      }
    },
    "ProductMdl" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "barcode" : {
          "type" : "string"
        },
        "category" : {
          "type" : "string",
          "enum" : [ "LACTATE", "FRUCTE", "CARNE", "DULCIURI" ]
        },
        "prices" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PriceMdl"
          }
        }
      }
    },
    "ProductResponseMdl" : {
      "type" : "object",
      "properties" : {
        "productMdl" : {
          "$ref" : "#/definitions/ProductMdl"
        },
        "bestPrice" : {
          "$ref" : "#/definitions/PriceMdl"
        },
        "otherPrices" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PriceMdl"
          }
        }
      }
    },
    "ProfileMdl" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "prices" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PriceMdl"
          }
        }
      }
    }
  }
}